refactor(tui): move chat viewport bridge wiring
